requirements.txt
pydantic-settings
python-dotenv
httpx
pyyaml
openai
pydantic>=2.0.0
websocket-client
Pillow
